+++ frontend/app/chart/page.tsx
@@ -8,7 +8,6 @@ import Sidebar from "../components/sidebar";
 export default async function Charts() {
   return (
     <>
-      <Sidebar></Sidebar>
       <div className="flex flex-col space-y-7 justify-center w-full h-full">
         <div className="mx-auto">
           <ExampleChart></ExampleChart>
