+++ frontend/app/components/testChart.tsx
@@ -368,7 +368,7 @@ export function FlowbiteDemo() {
 
   return (
     <>
-      <div className="p-4 bg-white border border-gray-200 rounded-lg shadow-sm 2xl:col-span-2 dark:border-gray-700 sm:p-6 dark:bg-gray-800">
+      <div className="p-4 bg-white border border-gray-200 rounded-lg shadow-sm w-full dark:border-gray-700 sm:p-6 dark:bg-gray-800">
         <div className="flex items-center justify-between mb-4">
           <div className="flex-shrink-0">
             <span className="text-xl font-bold leading-none text-gray-900 sm:text-2xl dark:text-white">
